+++ core/src/main/scala/com/evolutiongaming/kafka/flow/KeyFlowOf.scala
@@ -1,6 +1,7 @@
 package com.evolutiongaming.kafka.flow
 
 import cats.Monad
+import cats.effect.Resource
 import cats.effect.Sync
 import cats.syntax.all._
 import com.evolutiongaming.kafka.flow.timer.TimerFlowOf
@@ -13,7 +14,7 @@ trait KeyFlowOf[F[_], S, A] {
     context: KeyContext[F],
     persistence: Persistence[F, S, A],
     timers: TimerContext[F]
-  ): F[KeyFlow[F, A]]
+  ): Resource[F, KeyFlow[F, A]]
 
 }
 object KeyFlowOf {
@@ -24,8 +25,10 @@ object KeyFlowOf {
     tick: TickOption[F, S]
   ): KeyFlowOf[F, S, A] = { (context, persistence, timers) =>
     implicit val _context = context
-    timerFlowOf(context, persistence, timers) flatMap { timerFlow =>
-      KeyFlow.of(fold, tick, persistence, timerFlow)
+    Resource.liftF {
+      timerFlowOf(context, persistence, timers) flatMap { timerFlow =>
+        KeyFlow.of(fold, tick, persistence, timerFlow)
+      }
     }
   }